// hr-frontend/src/lib/api/generated/model/createUnitRelationshipRequest.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */

export interface CreateUnitRelationshipRequest {
  fromUnitId?: number;
  toUnitId: number;
  relationTypeId: number;
  startDate: string;
  endDate?: string;
  active?: boolean;
}