// hr-frontend/src/lib/api/generated/model/unitRelationship.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */

export interface UnitRelationship {
  id?: number;
  fromUnitId?: number;
  fromUnit?: string;
  relationTypeId?: number;
  type?: string;
  typeCode?: string;
  toUnitId?: number;
  toUnit?: string;
  startDate?: string;
  endDate?: string;
  status?: string;
}